Added refunds
This version integrates refunds. on initial testing I find refunds where the refund amount is not equal to the charge value, so we have partial refunds as well. Also we are using the max function - if we are already using row number = 1 then does it really matter if we use max or not?